src/main/java/com/jasmine/quizzy/BackgroundMusic.java
package com.jasmine.quizzy;

import javafx.scene.media.Media;
import javafx.scene.media.MediaPlayer;

import java.io.File;

public class BackgroundMusic {
    private MediaPlayer mediaPlayer;

    // Constructor for production code
    public BackgroundMusic(String resourcePath) {
        try {
            Media media = new Media(getClass().getResource("/backgroundMusic.mp3").toString());
            this.mediaPlayer = new MediaPlayer(media);
            mediaPlayer.setCycleCount(MediaPlayer.INDEFINITE);
        } catch (Exception e) {
            System.err.println("Error loading music: " + e.getMessage());
        }
    }

    // Constructor for testing
    public BackgroundMusic(MediaPlayer mediaPlayer) {
        this.mediaPlayer = mediaPlayer;
    }

    public void play() {
        if (mediaPlayer != null) mediaPlayer.play();
    }

    public void pause() {
        if (mediaPlayer != null) mediaPlayer.pause();
    }

    public void stop() {
        if (mediaPlayer != null) mediaPlayer.stop();
    }

    public void setVolume(double volume) {
        if (mediaPlayer != null) mediaPlayer.setVolume(volume);
    }

    public double getVolume() {
        return mediaPlayer != null ? mediaPlayer.getVolume() : 0.0;
    }
}
